sonarqube-msbuild-plugin/src/main/java/org/sonar/plugins/msbuild/MSBuildLanguage.java
/*
 * Sonar MSBuild Plugin :: Squid
 * Copyright (C) 2015-2018 SonarSource SA
 * mailto:info AT sonarsource DOT com
 *
 * This program is free software; you can redistribute it and/or
 * modify it under the terms of the GNU Lesser General Public
 * License as published by the Free Software Foundation; either
 * version 3 of the License, or (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
 * Lesser General Public License for more details.
 *
 * You should have received a copy of the GNU Lesser General Public License
 * along with this program; if not, write to the Free Software Foundation,
 * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 */
/*
 * Sonar MSBuild Plugin, open source software quality management tool.
 * Author(s) : Jorge Costa @ jmecsoftware.com
 * 
 * Sonar MSBuild Plugin is free software; you can redistribute it and/or
 * modify it under the terms of the GNU Lesser General Public
 * License as published by the Free Software Foundation; either
 * version 3 of the License, or (at your option) any later version.
 *
 * Sonar MSBuild Plugin is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
 * Lesser General Public License for more details.
 */
package org.sonar.plugins.msbuild;

import java.util.ArrayList;
import java.util.List;
import org.sonar.api.config.Configuration;
import org.sonar.api.resources.AbstractLanguage;

/**
 * {@inheritDoc}
 */
public class MSBuildLanguage extends AbstractLanguage {
  public static final String[] DEFAULT_SOURCE_SUFFIXES = {".vcxproj",".csproj",".fsproj",".msbuild",".props",".targets"};
  public static final String LANGUAGE_NAME = "MSBuild";    
  public static final String KEY = "msbuild";
  private final Configuration settings;
  
  /**
   * {@inheritDoc}
   * @param settings
   */
  public MSBuildLanguage(Configuration settings) {
    super(KEY, LANGUAGE_NAME);
    this.settings = settings;
  }
  
  /**
   * {@inheritDoc}
   * @return 
   */
  @Override
  public String[] getFileSuffixes() {
    String[] suffixes = filterEmptyStrings(settings.getStringArray(MSBuildPlugin.FILE_SUFFIXES_KEY));
    if (suffixes.length == 0) {
      suffixes = MSBuildLanguage.DEFAULT_SOURCE_SUFFIXES;
    }
    return suffixes;
  }

  private String[] filterEmptyStrings(String[] stringArray) {
    List<String> nonEmptyStrings = new ArrayList<>();
    for (String string : stringArray) {
      if (!"".equals(string.trim())) {
        nonEmptyStrings.add(string.trim());
      }
    }
    return nonEmptyStrings.toArray(new String[nonEmptyStrings.size()]);
  }  
}
